+++ trench/backends/base.py
@@ -2,7 +2,7 @@ import pyotp
 from abc import abstractmethod
 from typing import Any, Dict
 
-from trench.exceptions import MissingSourceFieldAttribute
+from trench.exceptions import MissingSourceFieldAttributeError
 from trench.settings import api_settings
 from trench.utils import create_otp_code, get_nested_attr_value
 
@@ -20,7 +20,7 @@ class AbstractMessageDispatcher:
         if self._FIELD_SOURCE_FIELD in conf:
             source = get_nested_attr_value(user, conf[self._FIELD_SOURCE_FIELD])
             if source is None:
-                raise MissingSourceFieldAttribute(  # pragma: no cover
+                raise MissingSourceFieldAttributeError(  # pragma: no cover
                     attribute_name=conf[self._FIELD_SOURCE_FIELD]
                 )
             self.to = source