+++ trench/serializers.py
@@ -1,40 +1,33 @@
-from typing import Any, Dict, Iterable
-
 from django.contrib.auth import authenticate, get_user_model
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Model
-from django.db.utils import DatabaseError
 from django.utils.translation import gettext as _
 
+from abc import abstractmethod
 from collections import OrderedDict
 from rest_framework.fields import CharField, ChoiceField
 from rest_framework.serializers import ModelSerializer, Serializer
+from typing import Any, Dict, Iterable
 
 from trench.exceptions import (
-    CodeInvalidOrExpired,
-    InvalidCodeValidationError,
-    InvalidTokenValidationError,
-    MFAMethodDoesNotExist,
-    MFAMethodNotRegisteredForUserValidationError,
-    MFANewPrimarySameAsOldValidationError,
-    MFANotEnabledValidationError,
-    MFAPrimaryMethodInactiveValidationError,
-    OTPCodeMissing,
-    RequiredFieldMissingValidationError,
-    RequiredFieldUpdateFailedValidationError,
-    UnauthenticatedValidationError,
-    UserAccountDisabledValidationError,
+    CodeInvalidOrExpiredError,
+    InvalidCodeError,
+    InvalidTokenError,
+    MFAMethodAlreadyActiveError,
+    MFAMethodDoesNotExistError,
+    MFANotEnabledError,
+    OTPCodeMissingError,
+    UnauthenticatedError,
+    UserAccountDisabledError,
 )
+from trench.query.get_mfa_method import get_mfa_method
 from trench.settings import api_settings
 from trench.utils import (
-    create_secret,
     get_mfa_handler,
     get_mfa_model,
-    get_nested_attr,
-    set_nested_attr,
     user_token_generator,
     validate_backup_code,
-    validate_code, get_method_config_by_name,
+    validate_code,
 )
 
 
@@ -48,16 +41,18 @@ ContextType = Dict[str, Any]
 
 
 def generate_model_serializer(name: str, model: Model, fields: Iterable[str]):
-    meta_subclass = type("Meta", (object,), {
-        "model": model,
-        "fields": fields,
-    })
-    return type(name, (ModelSerializer,), {
-        "Meta": meta_subclass
-    })
+    meta_subclass = type(
+        "Meta",
+        (object,),
+        {
+            "model": model,
+            "fields": fields,
+        },
+    )
+    return type(name, (ModelSerializer,), {"Meta": meta_subclass})
 
 
-class RequestValidator(Serializer):
+class RequestBodyValidator(Serializer):
     def update(self, instance, validated_data):
         pass
 
@@ -65,26 +60,87 @@ class RequestValidator(Serializer):
         pass
 
 
+class ProtectedActionValidator(RequestBodyValidator):
+    code = CharField(required=True)
+
+    @staticmethod
+    @abstractmethod
+    def _get_validation_method_name() -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def _validate_mfa_method(mfa: MFAMethod):
+        pass
+
+    def __init__(self, mfa_method_name: str, user: User, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._user = user
+        self._mfa_method_name = mfa_method_name
+
+    def validate_code(self, value: str) -> str:
+        if not value:
+            raise OTPCodeMissingError()
+
+        mfa = get_mfa_method(user_id=self._user.id, name=self._mfa_method_name)
+        self._validate_mfa_method(mfa)
+
+        validated_backup_code = validate_backup_code(value, mfa.backup_codes)
+
+        handler = get_mfa_handler(mfa)
+        validation_method = getattr(handler, self._get_validation_method_name())
+        if validation_method(value):
+            return value
+
+        if validated_backup_code:
+            mfa.remove_backup_code(validated_backup_code)
+            return value
+
+        raise CodeInvalidOrExpiredError()
+
+
+class MFAMethodDeactivationValidator(ProtectedActionValidator):
+    @staticmethod
+    def _get_validation_method_name() -> str:
+        return "validate_code"
+
+    @staticmethod
+    def _validate_mfa_method(mfa: MFAMethod):
+        if not mfa.is_active:
+            raise MFANotEnabledError()
+
+
+class MFAMethodActivationConfirmationValidator(ProtectedActionValidator):
+    @staticmethod
+    def _get_validation_method_name() -> str:
+        return "validate_confirmation_code"
+
+    @staticmethod
+    def _validate_mfa_method(mfa: MFAMethod):
+        if mfa.is_active:
+            raise MFAMethodAlreadyActiveError()
+
+
 class ProtectedActionSerializer(Serializer):
     requires_mfa_code = None
-    handler_validation_method = "validate_code"
+    _HANDLER_VALIDATION_METHOD = "validate_code"
 
     code = CharField(required=False)
 
     def _validate_code(self, value: str) -> str:
         if not value:
-            raise OTPCodeMissing()
+            raise OTPCodeMissingError()
 
         obj = self.context["obj"]
         validated_backup_code = validate_backup_code(value, obj.backup_codes)
         handler = get_mfa_handler(obj)
-        validate_method = getattr(handler, self.handler_validation_method)
+        validate_method = getattr(handler, self._HANDLER_VALIDATION_METHOD)
         if validate_method(value):
             return value
         if validated_backup_code:
             obj.remove_backup_code(validated_backup_code)
             return value
-        raise CodeInvalidOrExpired()
+        raise CodeInvalidOrExpiredError()
 
     def validate(self, data):
         if self.requires_mfa_code:
@@ -99,64 +155,21 @@ class ProtectedActionSerializer(Serializer):
         pass
 
 
-class RequestMFAMethodActivationConfirmSerializer(ProtectedActionSerializer):
-    requires_mfa_code = True
-    handler_validation_method = "validate_confirmation_code"
-
-
-class RequestMFAMethodDeactivationSerializer(ProtectedActionSerializer):
-    requires_mfa_code = api_settings.CONFIRM_DISABLE_WITH_CODE
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        context = kwargs["context"]
-        self.user = context["request"].user
-
-        is_current_method_primary = (
-            MFAMethod.objects.filter(user=self.user, name=context["name"])
-            .values_list("is_primary", flat=True)
-            .first()
-        )
-
-        self.users_active_methods_count = MFAMethod.objects.filter(
-            user=self.user, is_active=True
-        ).count()
-        if is_current_method_primary and self.users_active_methods_count > 2:
-            self.fields["new_primary_method"] = CharField(max_length=255, required=True)
-        else:
-            self.new_method = None
-
-    def validate_new_primary_method(self, value):
-        method_to_deactivate = self.context.get("name")
-
-        if method_to_deactivate == value:
-            raise MFANewPrimarySameAsOldValidationError()
-
-        try:
-            self.new_method = MFAMethod.objects.get(user=self.user, name=value)
-        except MFAMethod.DoesNotExist:
-            raise MFAMethodNotRegisteredForUserValidationError()
-        if not self.new_method.is_active:
-            raise MFAPrimaryMethodInactiveValidationError()
-        return value
-
-
 class RequestMFAMethodBackupCodesRegenerationSerializer(ProtectedActionSerializer):
     requires_mfa_code = api_settings.CONFIRM_BACKUP_CODES_REGENERATION_WITH_CODE  # noqa
 
 
-class RequestMFAMethodCodeSerializer(RequestValidator):
+class RequestMFAMethodCodeSerializer(RequestBodyValidator):
     method = CharField(max_length=255, required=False)
 
     @staticmethod
     def validate_method(value: str) -> str:
         if value and value not in api_settings.MFA_METHODS:
-            raise MFAMethodDoesNotExist()
+            raise MFAMethodDoesNotExistError()
         return value
 
 
-class LoginSerializer(RequestValidator):
+class LoginSerializer(RequestBodyValidator):
     """
     Validates user's credentials.
     """
@@ -176,15 +189,15 @@ class LoginSerializer(RequestValidator):
         )
 
         if not getattr(self.user, api_settings.USER_ACTIVE_FIELD, True):
-            raise UserAccountDisabledValidationError()
+            raise UserAccountDisabledError()
 
         if not self.user:
-            raise UnauthenticatedValidationError()
+            raise UnauthenticatedError()
 
         return {}
 
 
-class CodeLoginSerializer(RequestValidator):
+class CodeLoginSerializer(RequestBodyValidator):
     """
     Validates given token and OTP code.
     """
@@ -205,7 +218,7 @@ class CodeLoginSerializer(RequestValidator):
 
         self.user = user_token_generator.check_token(user=None, token=ephemeral_token)
         if not self.user:
-            raise InvalidTokenValidationError()
+            raise InvalidTokenError()
 
         for auth_method in self.user.mfa_methods.filter(is_active=True):
             validated_backup_code = validate_backup_code(
@@ -217,7 +230,7 @@ class CodeLoginSerializer(RequestValidator):
             if validated_backup_code:
                 auth_method.remove_backup_code(validated_backup_code)
                 return attrs
-        raise InvalidCodeValidationError()
+        raise InvalidCodeError()
 
 
 class UserMFAMethodSerializer(ModelSerializer):
@@ -230,7 +243,7 @@ class UserMFAMethodSerializer(ModelSerializer):
         fields = ("name", "is_primary")
 
 
-class ChangePrimaryMethodSerializer(RequestValidator):
+class ChangePrimaryMethodSerializer(RequestBodyValidator):
     """
     Serializes request to change default authentication method.
     """
@@ -246,14 +259,14 @@ class ChangePrimaryMethodSerializer(RequestValidator):
                 is_active=True,
             )
         except ObjectDoesNotExist:
-            raise MFANotEnabledValidationError()
+            raise MFANotEnabledError()
         try:
             new_primary_method = user.mfa_methods.get(
                 name=attrs.get("method"),
                 is_active=True,
             )
         except ObjectDoesNotExist:
-            raise MFAMethodDoesNotExist()
+            raise MFAMethodDoesNotExistError()
         code = attrs.get("code")
         validated_backup_code = validate_backup_code(
             code,
@@ -269,7 +282,7 @@ class ChangePrimaryMethodSerializer(RequestValidator):
             current_method.remove_backup_code(validated_backup_code)
             return attrs
         else:
-            raise InvalidCodeValidationError()
+            raise InvalidCodeError()
 
     def save(self):
         new_method = self.validated_data.get("new_method")