+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Used :code:`black` tool for auto code formatting
 * Type hits added
 * Backwards-compatible code refactoring and cleanup for better maintenance experience
+* Deactivation of primary MFA method would raise an error instead of randomly selecting new primary method from all active ones
 
 
 0.2.3 (2019-12-16)